% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicatorMetadata.R
\name{load_indicatorMetadata}
\alias{load_indicatorMetadata}
\title{Load Indicator Metadata (Internal Use)}
\usage{
load_indicatorMetadata()
}
\value{
A data frame containing metadata for indicators that are available and verified.
}
\description{
This function retrieves and returns the metadata for all indicator data sets included in the IA package. The metadata provides detailed information such as indicator names, associated data file names, and instructions for plotting and data transformation. Note that this function is intended for internal package use only. For a more user-friendly and structured overview of the metadata, use `load_indicatorSummary()`.
}
\examples{
load_indicatorMetadata()

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
