% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicatorSummary.R
\name{load_indicatorSummary}
\alias{load_indicatorSummary}
\title{Load Metadata for Indicators}
\usage{
load_indicatorSummary()
}
\value{
A data frame containing the metadata of indicators, including information such as indicator name, description, data sources, and other relevant attributes.
}
\description{
This function loads all the metadata of the indicators presented in currently available in the IA package.
}
\details{
The metadata is sourced from the supplementary material from the *IPBES Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature* (Table 1, section "Summary of indicators").
}
\examples{
# Load the metadata for the indicators
indicator_metadata <- load_indicatorSummary()
head(indicator_metadata)

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
