% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absoluteValue.R
\name{plot_absoluteValue}
\alias{plot_absoluteValue}
\title{Print Absolute values of Indicator over Time}
\usage{
plot_absoluteValue(indicatorFileName)
}
\arguments{
\item{The}{filename of the indicator that should be displayed}
}
\value{
figure plot with x-axes "year"  and y-axes "raw data"
}
\description{
Print Absolute values of Indicator over Time
}
\examples{
plot_absoluteValue("Aboveground_biomass_PgC.csv")
}
