% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_changeSince1970.R
\name{plot_changeSince1970}
\alias{plot_changeSince1970}
\title{Plot Relative Change of Indicator Over Time Since 1970}
\usage{
plot_changeSince1970(indicatorName)
}
\arguments{
\item{indicatorName}{A character string representing the full name of the indicator to be plotted. Use `print_indicatorNames()` to view a list of available indicators.}
}
\value{
A ggplot object showing the relative change of the indicator over time. The x-axis represents the year, and the y-axis represents the rescaled values, with the first available year since 1970 set to 100.
}
\description{
This function generates a plot showing the relative change of a given indicator using 1970 onward. The years are displayed on the x-axis, and the rescaled indicator values (relative to the first available year after 1970) are shown on the y-axis. For details on the rescaling procedure, refer to `rescale_to_1970()`.
}
\examples{
# Example: Plot the relative change since 1970 for the Aboveground biomass indicator
plot_changeSince1970("Aboveground_biomass_PgC.csv")

}
