% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_indicatorFileNames.R
\name{print_indicatorFileNames}
\alias{print_indicatorFileNames}
\title{Display Available Indicator File Names}
\usage{
print_indicatorFileNames()
}
\value{
A sorted vector of unique file names for the available indicator data sets.
}
\description{
This function prints the file names of all available indicator data sets in the IA package. These files are located in the `inst/Global_IA_Data` directory. The function filters out superseded indicators and those without valid file names, displaying only the active and available ones.
}
\examples{
# Example: Print the file names of available indicators
print_indicatorFileNames()

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
