% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_indicatorNames.R
\name{print_indicatorNames}
\alias{print_indicatorNames}
\title{Display Long Names of Available Indicators}
\usage{
print_indicatorNames()
}
\value{
A sorted list of unique long names of available indicators.
}
\description{
This function prints the long names of all available indicator data sets included in the IA package.
}
\examples{
# Example: Print the long names of available indicators
print_indicatorNames()

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
