% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_to_1970.R
\name{rescale_to_1970}
\alias{rescale_to_1970}
\title{Rescale Indicator Values Since 1970}
\usage{
rescale_to_1970(v, y, gt = 5)
}
\arguments{
\item{v}{Numeric vector representing the indicator values.}

\item{y}{Numeric vector representing the years associated with the indicator values.}

\item{gt}{Integer threshold for determining whether to use a GAM or linear regression for fitting (default is 5).}
}
\value{
A data frame with rescaled values, including the original data, fitted values, SEs, and the rescaled values relative to 1970.
}
\description{
This function trims off data prior to 1970 and rescales the remaining values as a percentage of the first available value after 1970.
It fits a line or curve to the data and uses the fitted value from the first year for rescaling, which is more reliable for noisy or fluctuating time series.
}
\details{
- For n = 2, a straight line is used with no standard errors (SEs).
- For 2 < n < gt, a linear regression is used to estimate fits, and SEs are returned with the fits.
- For n >= gt, a generalized additive model (GAM) is used. If n <= 10, the basis dimension `k` is set to 3; otherwise, the default is used. SEs are returned with the fits.
- For indicators where "High is better", the all-gone value is needed. For "Low is better", the pristine value is required.
}
