% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_to_percent.R
\name{rescale_to_percent}
\alias{rescale_to_percent}
\title{Rescale an Indicator to Percentage}
\usage{
rescale_to_percent(v, y, hl, pv, agv, gt = 5)
}
\arguments{
\item{v}{A numeric vector representing the raw values of the indicator.}

\item{y}{A numeric vector representing the years corresponding to the indicator values.}

\item{hl}{A string indicating whether "High" or "Low" values are better ("High" for indicators where higher values are better, "Low" for indicators where lower values are better).}

\item{pv}{Numeric value representing the pristine value of the indicator (the value in an intact world).}

\item{agv}{Numeric value representing the all-gone value of the indicator (the value in a fully degraded world).}

\item{gt}{An integer threshold for the number of data points required to use a generalized additive model (GAM). The default is 5.}
}
\value{
A data frame with the original values, rescaled percentages, and fitted percentages over time.
}
\description{
This function rescales an indicator to a percentage scale for indicators where the values of an intact (pristine) world and a fully-degraded world can be specified.
It calculates how far each time point is from pristine (100%) to fully degraded (0%).
If the indicator is not suitable for rescaling, the function will return a vector of NAs of the same length as the indicator.
}
\examples{
# Example usage:
# rescale_to_percent(v = indicator_values, y = years, hl = "High", pv = pristine_value, agv = all_gone_value)

}
