% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicatorData.R
\name{load_indicatorData}
\alias{load_indicatorData}
\title{Load Annual Global Data for Indicators}
\usage{
load_indicatorData(dataset.name)
}
\value{
A data frame containing year and global indicator value
}
\description{
This function loads a the global annual indicator data called by its name
}
\details{
The data is sourced from the supplementary material from the *IPBES Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
\examples{
# Load the metadata for the indicators
dat <- load_indicatorData("Aboveground_biomass_PgC.csv")
head(dat)
plot(dat)

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
