% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_indicatorSummary.R
\name{load_indicatorSummary}
\alias{load_indicatorSummary}
\title{Load Comprehensive Summary of Indicator Metadata}
\usage{
load_indicatorSummary()
}
\value{
A data frame containing comprehensive metadata for the available indicators, including details on the indicator's name, data sources, time span, EBV class, and other relevant attributes.
}
\description{
This function loads a detailed overview of the metadata for indicators available in the IA package. The summary includes key information such as the indicator name, data sources, time span covered, EBV class, and other relevant attributes.
}
\examples{
# Load the comprehensive metadata for the indicators
Tab1.dat <- load_indicatorSummary()
head(Tab1.dat)

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
