% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absoluteValue.R
\name{plot_absoluteValue}
\alias{plot_absoluteValue}
\title{Plot Absolute Values of Indicator Over Time}
\usage{
plot_absoluteValue(indicatorName)
}
\arguments{
\item{indicatorName}{A character string representing the full name of the indicator to be plotted. To view a list of available indicators, use the `print_indicatorNames()` function.}
}
\value{
A ggplot figure with "Year" on the x-axis and "Raw value" on the y-axis, representing the absolute values of the selected indicator.
}
\description{
This function generates a plot showing the absolute values of a given indicator over time, with years on the x-axis and raw indicator values on the y-axis.
}
\examples{
# Example: Plot the absolute values of the Aboveground biomass indicator
plot_absoluteValue("Aboveground_biomass_PgC.csv")

}
