% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_indicatorFileNames.R
\name{print_indicatorFileNames}
\alias{print_indicatorFileNames}
\title{Display Names of Indicators}
\usage{
print_indicatorFileNames()
}
\description{
This function prints the file names of all indicator data sets available in the IA package
}
\examples{
print_indicatorFileNames()

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
