% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_indicatorNames.R
\name{print_indicatorNames}
\alias{print_indicatorNames}
\title{Display Names of Indicators}
\usage{
print_indicatorNames()
}
\description{
This function prints the long names of all indicator data sets available in the IA package
}
\examples{
print_indicatorNames()

}
\references{
IPBES (2019). *Global Assessment Chapter 2.2 – Supplementary Material: Indicators of Status & Trends in Nature*.
}
